requirements.txt
python-telegram-bot==20.7
transformers==4.36.0
torch==2.1.0
datasets==2.16.0
accelerate==0.25.0
sentence-transformers==2.2.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
asyncio-mqtt==0.16.1
schedule==1.2.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
pandas==2.1.4
openpyxl==3.1.2 